Refactor model configuration HTTP routes
